Compiler warnings
git-svn-id: http://svn.ta3d.org/yuni/trunk@1614 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/directory/info/platform.cpp
@@ -214,7 +214,7 @@ namespace Directory
 		struct stat s;
 		# else
 		struct _wfinddatai64_t data;
-		long h;
+		intptr_t h;
 		bool callNext;
 		# endif
